pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gcp-auditor"
version = "0.6.0"
description = "Open-source GCP security scanner. 30 checks mapped to ISO 27001, SOC 2, and CIS GCP Benchmarks, each with a ready-to-use fix."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Abdullah Kamil" },
]
keywords = ["gcp", "google-cloud", "cloud", "audit", "security", "compliance", "iso27001", "soc2", "cis", "scanner", "devops", "terraform"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: System :: Systems Administration",
    "Typing :: Typed",
]
dependencies = [
    "google-auth>=2.25.0",
    "google-api-python-client>=2.100.0",
    "google-cloud-storage>=2.10.0",
    "typer>=0.15.0",
    "rich>=13.9.0",
    "pydantic>=2.10.0",
    "jinja2>=3.1.6",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-PyYAML>=6.0",
]

[project.scripts]
gcp-auditor = "cloud_audit.cli:app"

[project.urls]
Homepage = "https://github.com/abdullahkamil/gcp-auditor"
Documentation = "https://github.com/abdullahkamil/gcp-auditor#readme"
Repository = "https://github.com/abdullahkamil/gcp-auditor"
Issues = "https://github.com/abdullahkamil/gcp-auditor/issues"
Changelog = "https://github.com/abdullahkamil/gcp-auditor/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["src/cloud_audit"]

[tool.ruff]
target-version = "py310"
line-length = 120
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "S", "B", "A", "C4", "SIM", "TCH"]
ignore = ["S101"]

[tool.ruff.lint.per-file-ignores]
"src/cloud_audit/providers/*/checks/*.py" = ["E501", "S112"]
"src/cloud_audit/cli.py" = ["TC003"]
"tests/**" = ["TC003"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
